Make this compile after previous commit.
git-svn-id: https://www.soclib.fr/svn/trunk/soclib@2268 2cf4c1f0-a32b-0410-9068-d80f8be93c97

+++ soclib/module/internal_component/vci_vcache_wrapper/caba/source/src/vci_vcache_wrapper.cpp
@@ -1939,9 +1939,9 @@ r_cpt_dirty_bit_updt++;
                             }
                             else                    // SC request accepted
                             {
-                                uint32_t    cache_sc_way;
-                                uint32_t    cache_sc_set;
-                                uint32_t    cache_sc_word;
+                                size_t      cache_sc_way;
+                                size_t      cache_sc_set;
+                                size_t      cache_sc_word;
                                 uint32_t    cache_sc_rdata;
                                 bool        sc_hit;
